+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
     * `Piwik\Log::setLogLevel()` and `Piwik\Log::getLogLevel()`: log levels are now handled through DI configuration
     * `Piwik\Tracker::getDatetimeFromTimestamp()`: use `Piwik\Date::getDatetimeFromTimestamp()` instead
     * `Piwik\Menu\MenuAbstract::add()`: use `Piwik\Menu\MenuAbstract::addItem()` instead
+    * `Piwik\Metrics\Formatter::getCurrencyList()`: use `Piwik\Intl\Data\Provider\CurrencyDataProvider::getCurrencyList()` instead
 
 ## Piwik 2.14.0